Create WhenWillAntDie.java
We have a wooden plank of the length A units. Some ants are walking on the plank, each ant moves with speed 1 unit per second. Some of the ants move to the left, the other move to the right.
When two ants moving in two different directions meet at some point, they change their directions and continue moving again. Assume changing directions doesn't take any additional time.

When an ant reaches one end of the plank at a time t, it falls out of the plank imediately.

Given an integer A and two integer arrays B signifying left going ants and C signifying right going ants, the positions of the ants moving to the left and the right. Return the moment when the last ant(s) fall out of the plank.